Finish graph representation, set up test, initial test for parser

+++ src/main/java/graph/Arc.java
@@ -1,15 +1,13 @@
 package graph;
 
-import org.jgrapht.graph.DefaultWeightedEdge;
-
 /**
  * Created by nateeo on 28/07/17.
  */
 public class Arc {
     WeightedVertex _source;
     WeightedVertex _destination;
-    int _weight;
-    public Arc(WeightedVertex source, WeightedVertex destination, int weight) {
+    double _weight;
+    public Arc(WeightedVertex source, WeightedVertex destination, double weight) {
         _source = source;
         _destination = destination;
         _weight = weight;
@@ -23,7 +21,7 @@ public class Arc {
         return _destination;
     }
 
-    public int getWeight() {
+    public double getWeight() {
         return _weight;
     }
 }
